Modernize wrappers for AmpInfoTable and ExposureCatalog.
