+++ include/lsst/afw/table/python/catalog.h
@@ -125,7 +125,7 @@ class object in Python.
 @param[in] name   Name prefix of the record type, e.g. "Base" or "Simple".
 @param[in] isBase Whether this instantiation is only being used as a base class (used to set the class name).
 */
-// TODO: remove once all catalogs have been rewrapped with WrapperCollection
+// TODO: remove once PeakCatalog has been rewrapped with WrapperCollection
 template <typename Record>
 PyCatalog<Record> declareCatalog(pybind11::module &mod, std::string const &name, bool isBase = false) {
     namespace py = pybind11;
